feat(dv): language server and monaco editor implementations
suggestions after initial review:
- consolidation/removal of websocket code
- various removals of dead or not ready parts of the code
- removal of the materialization completion as we're not ready to call
that out for use yet
