+++ app/dv/src/main/resources/io/syndesis/dv/lsp/completion/providers/items/tableBodyOptionsCompletionItems.json
@@ -43,34 +43,6 @@
             "textEdit": "",
             "additionalTextEdits": "",
             "commitCharacters": ""
-        },
-        {
-            "label": "MATERIALIZED_TABLE",
-            "kind": "SNIPPET",
-            "detail": "",
-            "documentation": "If this view is being materialized to a external database, this defines the name of the table that is being materialized to.",
-            "deprecated": "false",
-            "preselect": "false",
-            "sortText": "false",
-            "insertText": "MATERIALIZED_TABLE '${1:tableName}'",
-            "insertTextFormat": "SNIPPET",
-            "textEdit": "",
-            "additionalTextEdits": "",
-            "commitCharacters": ""
-        },
-        {
-            "label": "MATERIALIZED",
-            "kind": "SNIPPET",
-            "detail": "",
-            "documentation": "Defines if a table is materialized ",
-            "deprecated": "false",
-            "preselect": "false",
-            "sortText": "false",
-            "insertText": "MATERIALIZED '${1:boolean}'",
-            "insertTextFormat": "SNIPPET",
-            "textEdit": "",
-            "additionalTextEdits": "",
-            "commitCharacters": ""
         }
     ]
 }
